add: split to_ts expr_stmt
